+++ api/src/reportcreator_api/tests/mock.py
@@ -243,18 +243,21 @@ def create_project(project_type=None, members=[], report_data={}, findings_kwarg
     for finding_kwargs in findings_kwargs if findings_kwargs is not None else [{}] * 3:
         create_finding(project=project, **finding_kwargs)
     
+    if notes_kwargs is not None:
+        # Delete default notes
+        project.notes.all().delete()
     for note_kwargs in notes_kwargs if notes_kwargs is not None else [{}] * 3:
         create_projectnotebookpage(project=project, **note_kwargs)
 
     for idx, image_kwargs in enumerate(images_kwargs if images_kwargs is not None else [{}] * 2):
         UploadedImage.objects.create(linked_object=project, **{
             'name': f'file{idx}.png', 
-            'file': SimpleUploadedFile(name=f'file{idx}.png', content=create_png_file())
+            'file': SimpleUploadedFile(name=f'file{idx}.png', content=image_kwargs.pop('content', create_png_file()))
         } | image_kwargs)
     for idx, file_kwargs in enumerate(files_kwargs if files_kwargs is not None else [{}] * 2):
         UploadedProjectFile.objects.create(linked_object=project, **{
             'name': f'file{idx}.pdf', 
-            'file': SimpleUploadedFile(name=f'file{idx}.pdf', content=f'%PDF-1.3{idx}'.encode())
+            'file': SimpleUploadedFile(name=f'file{idx}.pdf', content=file_kwargs.pop('content', f'%PDF-1.3{idx}'.encode()))
         } | file_kwargs)
 
     return project