+++ frontend/src/components/Btn/Import.vue
@@ -1,5 +1,6 @@
 <template>
   <s-btn-icon
+    v-if="buttonVariant === 'icon'"
     @click="fileInput.click()"
     :loading="props.loading || importInProgress"
     :disabled="props.disabled"
@@ -18,6 +19,28 @@
 
     <s-tooltip activator="parent" location="bottom" text="Import from file" />
   </s-btn-icon>
+  <v-list-item
+    v-else
+    link
+    @click="fileInput.click()"
+    :disabled="props.disabled"
+  >
+    <template #prepend>
+      <v-progress-circular v-if="props.loading || importInProgress" indeterminate size="24" />
+      <v-icon v-else icon="mdi-upload" />
+    </template>
+    <template #default>
+      <v-list-item-title>Import</v-list-item-title>
+      <input
+        ref="fileInput"
+        type="file"
+        accept=".tar.gz"
+        @change="performImport(($event.target as HTMLInputElement)?.files)"
+        class="d-none"
+        :disabled="disabled || props.loading || importInProgress"
+      />
+    </template>
+  </v-list-item>    
 </template>
 
 <script setup lang="ts">
@@ -25,9 +48,11 @@ const props = withDefaults(defineProps<{
   import: (file: File) => Promise<void>;
   loading?: boolean;
   disabled?: boolean;
+  buttonVariant: 'icon' | 'list-item';
 }>(), {
   loading: false,
-  disabled: false
+  disabled: false,
+  buttonVariant: 'icon',
 });
 
 const importInProgress = ref(false);