+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Next
+* Define initial note structure for projects in designs
+* Allow exporting and importing notes
+
+
 ## v2024.8 - 2024-01-23
 * Diff-view for version history
 * Set form fields readonly instead of disabled